uint8_t

+++ afsk/main.c
@@ -118,7 +118,7 @@ struct SensorData read_sensor_data(struct SensorConfig sensor) {
     delay(1); // Max 12-bit conversion time is 586us per sample	
    //int16_t value = (int16_t)((wiringPiI2CRead(sensor.fd) << 8 ) | wiringPiI2CRead (sensor.fd));
   //int16_t value = (( ((int16_t)byte) << 8 ) | (int16_t)wiringPiI2CRead (sensor.fd));
-    int16_t value = (int16_t)(( ((uint16_t)byte) << 8 ) | (uint16_t)wiringPiI2CRead (sensor.fd));
+    int16_t value = (int16_t)(( ((uint8_t)byte) << 8 ) | (uint8_t)wiringPiI2CRead (sensor.fd));
     float value1 = (float) value;
     data.current  = value1 / (float)sensor.currentDivider;
     printf("********* value0 %d value1 %f value %d data.current %f \n", value0, value1, value, data.current);
